+++ models/productDao.js
@@ -1,4 +1,8 @@
 const { appDataSource } = require("./dataSource");
+const {
+  orderSet,
+  makeProductQueryBuilders,
+} = require("../models/productQueryBuilder");
 
 const getProductDetail = async (productId) => {
   return await appDataSource.query(
@@ -45,8 +49,18 @@ const getProductDetail = async (productId) => {
   );
 };
 
-const getProductList = async (whereClause, orderbyClause) => {
-  return await appDataSource.query(
+const getProductList = async (params) => {
+  const {
+    limit = 10,
+    offset = 0,
+    sortMethod = "created_at",
+    ...filterOptions
+  } = params;
+
+  const whereClause = makeProductQueryBuilders(filterOptions);
+  const orderbyClause = orderSet[sortMethod];
+
+  const result = await appDataSource.query(
     `
         SELECT 
             sub_categories.name AS subcategory,
@@ -63,6 +77,7 @@ const getProductList = async (whereClause, orderbyClause) => {
         ORDER BY ${orderbyClause}
         `
   );
+  return result;
 };
 
 module.exports = {
